prepared Regent for mixin-ifying. added plugin to regent getDb() function
