+++ lib/db/relational/mysql/connection.js
@@ -2,6 +2,7 @@
  * @author Steven Jimenez <steven@stevethedev.com>
  */
 'use strict';
+const Database        = require('regent/lib/db/database');
 
 // External MySQL Connection Dependency
 const mysql           = require('mysql');
@@ -202,3 +203,6 @@ function getConfig() {
 }
 
 module.exports = MysqlDb;
+
+Database.registerDriver('MySQL', MysqlDb);
+
