+++ lib/db/relational/postgresql/connection.js
@@ -6,6 +6,8 @@
 // External PostgreSQL Connection
 const { Pool }        = require('pg');
 
+const Database        = require('regent/lib/db/database');
+
 const RelationalDb    = require('regent/lib/db/relational/connection');
 const QueryBuilder    = require(
     'regent/lib/db/relational/postgresql/query-builder'
@@ -160,5 +162,7 @@ function getConfig() {
         user                   : cfg.username,
     };
 }
+
+Database.registerDriver('PostgreSQL', PostgresDb);
 
 module.exports = PostgresDb;
